code/web-app/backend/requirements.txt
web3
gql
requests
python-dotenv
requests_toolbelt
asyncio
aiohttp
pymysql
flask
mysql-connector-python
flask-cors
